feat: move logic to get invalid rows from store to query
